Update registrar.lookup. More async wrap.
Replace registrar lookup with a namepace lookup. It was more of a placeholder anyway, but it was a little clunky and not really useful. By extension, add and get_by_namespace now accept a class for namespace.

setUpModule and tearDownModule do not support async, so provide non-async versions of 'clear', etc.