Refactor entity and component
In my enthusiasm, I did a dumb and completely forgot a pretty crucial part of ECS: Entity is only a UUID. Not a huge fix, but I added some unnecessary stuff. Removed the unnecessary and reworked. I did decide to have Entity store the the component IDs for now instead of reworking Registrar.